# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Shopify API",
        "description": "RESTful API for managing Shopify products and variants with GraphQL integration",
        "contact": {
            "name": "API Support",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/api/v1/shopify/import": {
            "post": {
                "tags": [
                    "Shopify"
                ],
                "summary": "Import products from Shopify",
                "description": "Import products from Shopify to local database with pagination support",
                "operationId": "b47fdf99fab727c63fef8f466d041157",
                "parameters": [
                    {
                        "name": "first",
                        "in": "query",
                        "description": "Number of products to import",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 250
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Pagination cursor for fetching next page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "update_existing",
                        "in": "query",
                        "description": "Whether to update existing products",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products imported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "imported": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "updated": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 35
                                        },
                                        "hasNextPage": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "endCursor": {
                                            "type": "string",
                                            "example": "eyJpZCI6IjEyMzQ1Njc4OTAifQ=="
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Import completed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error - Shopify API error"
                    }
                }
            }
        },
        "/api/v1/shopify/export/{product}": {
            "post": {
                "tags": [
                    "Shopify"
                ],
                "summary": "Export specific product to Shopify",
                "description": "Export a local product to Shopify (creates new or updates existing)",
                "operationId": "178968bf4360a8ef88da9a220b3e506a",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID to export",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "force_create",
                        "in": "query",
                        "description": "Force create new product even if shopify_product_id exists",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product exported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product exported successfully"
                                        },
                                        "shopify_product_id": {
                                            "type": "integer",
                                            "example": 987654321
                                        },
                                        "shopify_gid": {
                                            "type": "string",
                                            "example": "gid://shopify/Product/987654321"
                                        },
                                        "synced_variants": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error - Product missing required data"
                    },
                    "500": {
                        "description": "Internal Server Error - Shopify API error"
                    }
                }
            }
        },
        "/api/v1/shopify/export/bulk": {
            "post": {
                "tags": [
                    "Shopify"
                ],
                "summary": "Export multiple products to Shopify",
                "description": "Bulk export products to Shopify (operates in background)",
                "operationId": "51a9c7b12100304b1c8eb0df03d9fb18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_ids": {
                                        "description": "Array of product IDs to export",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3,
                                            4,
                                            5
                                        ]
                                    },
                                    "update_existing": {
                                        "description": "Update existing products in Shopify",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "skip_failed": {
                                        "description": "Continue processing if some products fail",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Bulk export started successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkOperationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid product IDs"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/shopify/sync/status": {
            "get": {
                "tags": [
                    "Shopify"
                ],
                "summary": "Get sync status overview",
                "description": "Get comprehensive sync status between local database and Shopify",
                "operationId": "ab11c3a7cc856327e068461f06b9ee04",
                "responses": {
                    "200": {
                        "description": "Sync status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Sync status",
                                            "type": "string",
                                            "enum": [
                                                "synced",
                                                "pending",
                                                "error"
                                            ],
                                            "example": "synced"
                                        },
                                        "last_sync_at": {
                                            "description": "Last sync timestamp",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-20T15:45:00Z"
                                        },
                                        "total_products": {
                                            "description": "Total products in system",
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "synced_products": {
                                            "description": "Products synced to Shopify",
                                            "type": "integer",
                                            "example": 145
                                        },
                                        "pending_products": {
                                            "description": "Products pending sync",
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "sync_errors": {
                                            "description": "Sync error messages",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "Product 123 sync failed"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/shopify/sync/validate": {
            "post": {
                "tags": [
                    "Shopify"
                ],
                "summary": "Validate Shopify credentials",
                "description": "Test Shopify API credentials and connection",
                "operationId": "60b241944a32e8e1f01a0103b9084193",
                "responses": {
                    "200": {
                        "description": "Credentials are valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "shop_name": {
                                            "type": "string",
                                            "example": "My Awesome Store"
                                        },
                                        "shop_domain": {
                                            "type": "string",
                                            "example": "mystore.myshopify.com"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shopify credentials are valid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid access token or store domain"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token authentication. Use the token returned from /api/v1/auth/login endpoint in the format: Bearer {your_token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}